+++ test/runtests.jl
@@ -22,6 +22,7 @@ const TEST_MODE = get(ENV, "TEST_MODE", "default")
         include("unit/test_sunny.jl")
         include("unit/test_vaspio.jl")
         include("unit/test_site_engine.jl")
+        include("unit/test_mfa_sampler.jl")
     end
     if TEST_MODE in ("default", "all", "aqua")
         include("aqua.jl")